requirements.txt
django==4.1.5
psycopg2-binary==2.9.*
python-dotenv==0.21.0
requests==2.28.2

